data_service/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
dnspython==2.4.2
exceptiongroup==1.1.3
fastapi==0.103.2
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
motor==3.3.1
mysql-connector-python==8.1.0
protobuf==4.21.12
pydantic==2.4.2
pydantic_core==2.10.1
pymongo==4.5.0
python-multipart==0.0.6
python-telegram-bot==20.6
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.0.6
uvicorn==0.23.2
Whoosh==2.7.4
